system_backend/requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
Brotli==1.1.0
celery==5.5.3
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.5
cryptography==45.0.4
cssselect2==0.8.0
defusedxml==0.7.1
Django==5.2.3
django-celery-beat==2.8.1
django-cors-headers==4.7.0
django-filter==25.1
django-js-asset==3.1.2
django-mptt==0.17.0
django-timezone-field==7.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
djoser==2.3.1
drf-yasg==1.21.10
fonttools==4.59.0
idna==3.10
inflection==0.5.1
kombu==5.5.4
Markdown==3.8.2
numpy==2.3.1
oauthlib==3.3.1
packaging==25.0
pandas==2.3.0
pillow==11.2.1
prompt_toolkit==3.0.51
pycparser==2.22
pydyf==0.11.0
PyJWT==2.9.0
pyphen==0.17.2
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python3-openid==3.2.0
pytz==2025.2
PyYAML==6.0.2
qrcode==8.2
reportlab==4.4.3
requests==2.32.4
requests-oauthlib==2.0.0
six==1.17.0
social-auth-app-django==5.4.3
social-auth-core==4.6.1
sqlparse==0.5.3
tinycss2==1.4.0
tinyhtml5==2.0.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
weasyprint==65.1
webencodings==0.5.1
xlrd==2.0.2
xlsxwriter==3.2.5
zopfli==0.2.3.post1
